Extracted the SearchPeople component logic into a new useSearchPeople custom hook
