+++ src/features/SearchPeople/index.js
@@ -1,12 +1,4 @@
-import { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { useQueryParameter } from "../../common/NavigationBar/SearchBar/queryParameter";
-import {
-    fetchSearchPeopleLoad,
-    selectSearchResults,
-    selectStatus
-} from "../People/peopleSlice";
-import { selectPage } from "../../common/Pagination/paginationSlice";
+import { useSearchPeople } from "./useSearchPeople";
 import PeopleTile from "../../common/Tiles/PeopleTile";
 import Pagination from "../../common/Pagination";
 import Loading from "../Actions/Loading";
@@ -16,24 +8,14 @@ import { PeopleList } from "../People/styled";
 import { SectionHeading } from "../../common/styled";
 
 const SearchPeople = () => {
-    const dispatch = useDispatch();
-    const query = useQueryParameter("search");
-
-    const status = useSelector(selectStatus);
-    const selectedPage = useSelector(selectPage);
-    const searchedResult = useSelector(selectSearchResults);
-
-    const searchedPeople = searchedResult.results;
-
-    const resultCurrentPage = searchedResult.page;
-    const resultTotalPages = searchedResult.total_pages;
-    const totalResults = searchedResult.total_results;
-
-    useEffect(() => {
-        if (query) {
-            dispatch(fetchSearchPeopleLoad(query, { page: selectedPage }));
-        }
-    }, [dispatch, query, selectedPage]);
+    const {
+        status,
+        query,
+        searchedPeople,
+        resultCurrentPage,
+        resultTotalPages,
+        totalResults
+    } = useSearchPeople();
 
     if (status === "error") {
         return <Error />;
